requirements.txt
blinker==1.9.0
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Limiter==3.9.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pillow==11.0.0
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
svgwrite==1.4.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
